+++ src/core/source_buffers/queued_source_buffer.ts
@@ -15,16 +15,25 @@
  */
 
 import {
+  fromEvent,
+  interval,
   Observable,
   Observer,
   Subject,
 } from "rxjs";
 import {
+  takeUntil,
+  tap,
+} from "rxjs/operators";
+import {
   ICustomSourceBuffer,
   tryToChangeSourceBufferType,
 } from "../../compat";
+import config from "../../config";
 import log from "../../log";
 import ICustomTimeRanges from "./time_ranges";
+
+const { SOURCE_BUFFER_FLUSHING_INTERVAL } = config;
 
 // Every QueuedSourceBuffer types
 export type IBufferType = "audio"|"video"|"text"|"image";
@@ -134,10 +143,12 @@ export default class QueuedSourceBuffer<T> {
    */
   private readonly _sourceBuffer : ICustomSourceBuffer<T>;
 
-  // Binded references to corresponding private methods.
-  // Used for binding/removing an event listener.
-  private readonly __onErrorEvent : (e : Event) => void;
-  private readonly __onUpdateEnd : () => void;
+  /**
+   * Subject triggered when this QueuedSourceBuffer is disposed.
+   * Helps to clean-up Observables created at its creation.
+   * @type {Subject}
+   */
+  private _destroy$ : Subject<void>;
 
   /**
    * Queue of awaited buffer orders.
@@ -192,6 +203,7 @@ export default class QueuedSourceBuffer<T> {
     codec : string,
     sourceBuffer : ICustomSourceBuffer<T>
   ) {
+    this._destroy$ = new Subject<void>();
     this.bufferType = bufferType;
     this._sourceBuffer = sourceBuffer;
     this._queue = [];
@@ -199,10 +211,25 @@ export default class QueuedSourceBuffer<T> {
     this._lastInitSegment = null;
     this._currentCodec = codec;
 
-    this.__onErrorEvent = this._onErrorEvent.bind(this);
-    this.__onUpdateEnd = this._onUpdateEnd.bind(this);
-    this._sourceBuffer.addEventListener("error", this.__onErrorEvent);
-    this._sourceBuffer.addEventListener("updateend", this.__onUpdateEnd);
+    // Some browsers (happened with firefox 66) sometimes "forget" to send us
+    // `update` or `updateend` events.
+    // In that case, we're completely unable to continue the queue here and
+    // stay locked in a waiting state.
+    // This interval is here to check at regular intervals if the underlying
+    // SourceBuffer is currently updating.
+    interval(SOURCE_BUFFER_FLUSHING_INTERVAL).pipe(
+      tap(() => this._flush()),
+      takeUntil(this._destroy$)
+    ).subscribe();
+
+    fromEvent(this._sourceBuffer, "error").pipe(
+      tap((err) => this._onErrorEvent(err)),
+      takeUntil(this._destroy$)
+    ).subscribe();
+    fromEvent(this._sourceBuffer, "updateend").pipe(
+      tap(() => this._flush()),
+      takeUntil(this._destroy$)
+    ).subscribe();
   }
 
   /**
@@ -227,6 +254,8 @@ export default class QueuedSourceBuffer<T> {
    * @returns {Observable}
    */
   public appendBuffer(infos : IAppendBufferInfos<T>) : Observable<unknown> {
+    log.debug("QSB: receiving order to push data to the SourceBuffer",
+      this.bufferType);
     return this._addToQueue({ type: SourceBufferAction.Append, value: infos });
   }
 
@@ -237,6 +266,8 @@ export default class QueuedSourceBuffer<T> {
    * @returns {Observable}
    */
   public removeBuffer(start : number, end : number) : Observable<unknown> {
+    log.debug("QSB: receiving order to remove data from the SourceBuffer",
+      this.bufferType);
     return this._addToQueue({ type: SourceBufferAction.Remove, value: { start, end } });
   }
 
@@ -256,8 +287,8 @@ export default class QueuedSourceBuffer<T> {
    * @private
    */
   public dispose() : void {
-    this._sourceBuffer.removeEventListener("error", this.__onErrorEvent);
-    this._sourceBuffer.removeEventListener("updateend", this.__onUpdateEnd);
+    this._destroy$.next();
+    this._destroy$.complete();
 
     if (this._currentOrder != null) {
       this._currentOrder.subject.complete();
@@ -284,14 +315,6 @@ export default class QueuedSourceBuffer<T> {
   }
 
   /**
-   * Callback used for the 'updateend' event, as a segment has been added/removed.
-   * @private
-   */
-  private _onUpdateEnd() : void {
-    this._flush();
-  }
-
-  /**
    * Callback used for the 'error' event from the SourceBuffer.
    * @private
    * @param {Event} error
@@ -308,7 +331,7 @@ export default class QueuedSourceBuffer<T> {
    * @private
    * @param {Error|Event} err
    */
-  private _onErrorEvent(err: Error|Event) : void {
+  private _onErrorEvent(err: unknown) : void {
     this._onError(
       err instanceof Error ?
         err :