+++ src/parsers/manifest/dash/parse_mpd.ts
@@ -15,12 +15,16 @@
  */
 
 import config from "../../../config";
+import arrayFind from "../../../utils/array_find";
 import idGenerator from "../../../utils/id_generator";
 import resolveURL, {
   normalizeBaseURL,
 } from "../../../utils/resolve_url";
 import { IParsedManifest } from "../types";
 import checkManifestIDs from "../utils/check_manifest_ids";
+import getClockOffset from "./get_clock_offset";
+import getHTTPUTCTimingURL from "./get_http_utc-timing_url";
+import getLastTimeReference from "./get_last_time_reference";
 import getPresentationLiveGap from "./get_presentation_live_gap";
 import {
   createMPDIntermediateRepresentation,
@@ -30,9 +34,17 @@ import {
   createPeriodIntermediateRepresentation,
   IPeriodIntermediateRepresentation,
 } from "./node_parsers/Period";
+import parseAvailabilityStartTime from "./parse_availability_start_time";
+import parseDuration from "./parse_duration";
 import parsePeriods from "./parse_periods";
 
 const generateManifestID = idGenerator();
+
+export interface IMPDParserArguments {
+  url : string; // URL of the manifest (post-redirection if one)
+  referenceDateTime? : number; // Default base time, in seconds
+  loadExternalClock: boolean; // If true, we might need to synchronize the clock
+}
 
 export type IParserResponse<T> =
   {
@@ -49,27 +61,27 @@ export type IParserResponse<T> =
 
 /**
  * @param {Element} root - The MPD root.
- * @param {string} url - The url where the MPD is located
+ * @param {Object} args
  * @returns {Object}
  */
 export default function parseMPD(
   root : Element,
-  uri : string
+  args : IMPDParserArguments
 ) : IParserResponse<IParsedManifest> {
   // Transform whole MPD into a parsed JS object representation
   const mpdIR = createMPDIntermediateRepresentation(root);
-  return loadExternalRessourcesAndParse(mpdIR, uri);
+  return loadExternalRessourcesAndParse(mpdIR, args);
 }
 
 /**
  * Checks if xlinks needs to be loaded before actually parsing the manifest.
  * @param {Object} mpdIR
- * @param {string} uri
+ * @param {Object} args
  * @returns {Object}
  */
 function loadExternalRessourcesAndParse(
   mpdIR : IMPDIntermediateRepresentation,
-  uri : string
+  args : IMPDParserArguments
 ) : IParserResponse<IParsedManifest> {
   const xlinksToLoad : Array<{ index : number; ressource : string }> = [];
   for (let i = 0; i < mpdIR.children.periods.length; i++) {
@@ -80,8 +92,7 @@ function loadExternalRessourcesAndParse(
   }
 
   if (xlinksToLoad.length === 0) {
-    const parsedManifest = parseCompleteIntermediateRepresentation(mpdIR, uri);
-    return { type: "done", value: parsedManifest };
+    return parseCompleteIntermediateRepresentation(mpdIR, args);
   }
 
   return {
@@ -114,7 +125,7 @@ function loadExternalRessourcesAndParse(
           // replace original "xlinked" periods by the real deal
           mpdIR.children.periods.splice(index, 1, ...periodsIR);
         }
-        return loadExternalRessourcesAndParse(mpdIR, uri);
+        return loadExternalRessourcesAndParse(mpdIR, args);
       },
     },
   };
@@ -123,26 +134,24 @@ function loadExternalRessourcesAndParse(
 /**
  * Parse the MPD intermediate representation into a regular Manifest.
  * @param {Object} mpdIR
- * @param {string} uri
+ * @param {Object} args
  * @returns {Object}
  */
 function parseCompleteIntermediateRepresentation(
   mpdIR : IMPDIntermediateRepresentation,
-  uri : string
-) : IParsedManifest {
+  args : IMPDParserArguments
+) : IParserResponse<IParsedManifest> {
   const {
     children: rootChildren,
     attributes: rootAttributes,
   } = mpdIR;
 
-  const baseURL = resolveURL(normalizeBaseURL(uri), rootChildren.baseURL);
+  const baseURL = resolveURL(normalizeBaseURL(args.url), rootChildren.baseURL);
+
+  const availabilityStartTime =
+    parseAvailabilityStartTime(rootAttributes, args.referenceDateTime);
 
   const isDynamic : boolean = rootAttributes.type === "dynamic";
-  const availabilityStartTime = (
-    rootAttributes.type === "static" ||
-    rootAttributes.availabilityStartTime == null
-  ) ?  0 : rootAttributes.availabilityStartTime;
-
   const parsedPeriods = parsePeriods(rootChildren.periods, {
     availabilityStartTime,
     duration: rootAttributes.duration,
@@ -150,23 +159,18 @@ function parseCompleteIntermediateRepresentation(
     baseURL,
   });
 
-  const duration : number|undefined = (() => {
-    if (rootAttributes.duration != null) {
-      return rootAttributes.duration;
-    }
-    if (isDynamic) {
-      return undefined;
-    }
-    if (parsedPeriods.length) {
-      const lastPeriod = parsedPeriods[parsedPeriods.length - 1];
-      if (lastPeriod.end != null) {
-        return lastPeriod.end;
-      } else if (lastPeriod.duration != null) {
-        return lastPeriod.start + lastPeriod.duration;
-      }
-    }
-    return undefined;
-  })();
+  const duration : number|undefined = parseDuration(rootAttributes, parsedPeriods);
+
+  const directTiming = arrayFind(rootChildren.utcTimings,
+    (utcTiming) =>
+      utcTiming.schemeIdUri === "urn:mpeg:dash:utc:direct:2014" &&
+      utcTiming.value != null
+  );
+
+  // second condition not needed but TS did not help there, even with a `is`
+  const clockOffsetFromDirectUTCTiming =
+    directTiming != null && directTiming.value != null ?
+      Date.now() - Date.parse(directTiming.value) : undefined;
 
   const parsedMPD : IParsedManifest = {
     availabilityStartTime,
@@ -177,10 +181,13 @@ function parseCompleteIntermediateRepresentation(
     periods: parsedPeriods,
     transportType: "dash",
     isLive: isDynamic,
-    uris: [uri, ...rootChildren.locations],
+    uris: [args.url, ...rootChildren.locations],
     suggestedPresentationDelay: rootAttributes.suggestedPresentationDelay != null ?
       rootAttributes.suggestedPresentationDelay :
       config.DEFAULT_SUGGESTED_PRESENTATION_DELAY.DASH,
+    clockOffset: clockOffsetFromDirectUTCTiming != null &&
+      !isNaN(clockOffsetFromDirectUTCTiming) ?
+        clockOffsetFromDirectUTCTiming : undefined,
   };
 
   // -- add optional fields --
@@ -198,7 +205,32 @@ function parseCompleteIntermediateRepresentation(
 
   checkManifestIDs(parsedMPD);
   if (parsedMPD.isLive) {
-    parsedMPD.presentationLiveGap = getPresentationLiveGap(parsedMPD);
+    const lastTimeReference = getLastTimeReference(parsedMPD);
+    if (
+      clockOffsetFromDirectUTCTiming == null &&
+      lastTimeReference == null && args.loadExternalClock
+    ) {
+      const UTCTimingHTTPURL = getHTTPUTCTimingURL(mpdIR);
+      if (UTCTimingHTTPURL != null && UTCTimingHTTPURL.length > 0) {
+        return {
+          type: "needs-ressources",
+          value: {
+            ressources: [UTCTimingHTTPURL],
+            continue: function continueParsingMPD(loadedRessources : string[]) {
+              if (loadedRessources.length !== 1) {
+                throw new Error("DASH parser: wrong number of loaded ressources.");
+              }
+              parsedMPD.clockOffset = getClockOffset(loadedRessources[0]);
+              parsedMPD.presentationLiveGap =
+                getPresentationLiveGap(parsedMPD, lastTimeReference);
+              return { type: "done", value: parsedMPD };
+            },
+          },
+        };
+      }
+    }
+    parsedMPD.presentationLiveGap = getPresentationLiveGap(parsedMPD, lastTimeReference);
   }
-  return parsedMPD;
+
+  return { type: "done", value: parsedMPD };
 }