+++ src/transports/types.ts
@@ -129,6 +129,14 @@ export type ILoaderObservable<T> = Observable<ILoaderEvent<T>>;
 export interface IManifestParserArguments<T, U> {
   response : ILoaderResponseValue<T>; // Response from the loader
   url : string; // URL originally requested
+  hasClockSynchronization : boolean; // If true, the current device is currently
+                                     // synchronized with the server's clock.
+                                     // If false, there may be a delay/advance
+                                     // between it and the client's clock.
+                                     // In the latter case, you might need to
+                                     // perform a synchronization step,
+                                     // depending on the type of
+                                     // Manifest.
 
   // allow the parser to load supplementary ressources (of type U)
   scheduleRequest : (request : () => Observable<U>) => Observable<U>;
@@ -267,16 +275,14 @@ interface IParsedKeySystem {
 }
 
 export interface ITransportOptions {
-  // every transports
-  segmentLoader? : CustomSegmentLoader;
+  aggressiveMode? : boolean;
+  keySystems? : (hex? : Uint8Array) => IParsedKeySystem[]; // TODO deprecate
   manifestLoader?: CustomManifestLoader;
-
-  // smooth only
-  suggestedPresentationDelay? : number;
+  minRepresentationBitrate? : number; // TODO deprecate
   referenceDateTime? : number;
-  minRepresentationBitrate? : number;
-  keySystems? : (hex? : Uint8Array) => IParsedKeySystem[];
   representationFilter? : IRepresentationFilter;
+  segmentLoader? : CustomSegmentLoader;
+  suggestedPresentationDelay? : number;
   supplementaryImageTracks? : ISupplementaryImageTrack[];
   supplementaryTextTracks? : ISupplementaryTextTrack[];
 }